+++ src/Services/Http/Engine.php
@@ -1,6 +1,8 @@
 <?php
 namespace Ebanx\Benjamin\Services\Http;
 
+use Ebanx\Benjamin\Facade;
+
 class Engine
 {
     /**
@@ -14,6 +16,14 @@ class Engine
     private $curlInfo;
 
     /**
+     * @var array;
+     */
+    private $userAgentInfo = [];
+    /**
+     * @var string;
+     */
+    private $formattedUserAgentInfo = '';
+    /**
      * @param String $method
      * @param String $url
      * @param array|object|boolean $data
@@ -24,6 +34,7 @@ class Engine
     private function sendRequest($method, $url, $data = false)
     {
         $curlHandler = curl_init();
+        curl_setopt($curlHandler, CURLOPT_HTTPHEADER, $this->formatUserAgentInfo());
 
         if ($method === 'POST') {
             curl_setopt($curlHandler, CURLOPT_POST, 1);
@@ -97,5 +108,29 @@ class Engine
     public function getInfo()
     {
         return $this->curlInfo;
+    }
+
+    public function addUserAgentInfo($userValue)
+    {
+        array_push($this->userAgentInfo, $userValue);
+    }
+
+    public function getUserAgentInfo()
+    {
+        return $this->userAgentInfo;
+    }
+
+    public function getFormattedUserAgentInfo()
+    {
+        return $this->formattedUserAgentInfo;
+    }
+
+    private function formatUserAgentInfo()
+    {
+        if (empty($this->formattedUserAgentInfo)) {
+            $formattedUserAgentInfo = ['X-Ebanx-Client-User-Agent: SDK-PHP/' . Facade::VERSION . ' ' . join(' ', $this->userAgentInfo)];
+            $this->formattedUserAgentInfo= $formattedUserAgentInfo;
+        }
+        return $this->formattedUserAgentInfo;
     }
 }