+++ tests/Unit/Services/Gateways/CreditCardTest.php
@@ -255,6 +255,137 @@ class CreditCardTest extends GatewayTestCase
         );
     }
 
+    public function testMinInstalmentsArgentina()
+    {
+        $country = Country::ARGENTINA;
+        $gateway = $this->setupGateway(1, new Config());
+        $value = CreditCardConfig::acquirerMinInstalmentValueForCurrency(Currency::localForCountry($country));
+
+        $paymentTerms = $gateway->getPaymentTermsForCountryAndValue($country, $value);
+        $this->assertEquals(
+            1,
+            count($paymentTerms)
+        );
+    }
+
+    public function testMinInstalmentsColombia()
+    {
+        $country = Country::COLOMBIA;
+        $gateway = $this->setupGateway(1, new Config());
+        $value = CreditCardConfig::acquirerMinInstalmentValueForCurrency(Currency::localForCountry($country));
+
+        $paymentTerms = $gateway->getPaymentTermsForCountryAndValue($country, $value);
+        $this->assertEquals(
+            1,
+            count($paymentTerms)
+        );
+    }
+
+    public function testInstalmentsArgentina()
+    {
+        $country = Country::ARGENTINA;
+        $usdToBrlRate = 1;
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config());
+        $totalInstalments = 6;
+        $instalmentsArray = $creditCard->getInstalmentsByCountry($country);
+        $expectedInstalmentsArray = array(
+            '1' => '1',
+            '2' => '2',
+            '3' => '3',
+            '6' => '6',
+            '9' => '9',
+            '12' => '12');
+
+        $this->assertEquals($totalInstalments, count($instalmentsArray));
+        $this->assertEquals($expectedInstalmentsArray, $instalmentsArray);
+    }
+
+    public function testInstalmentsBrazil()
+    {
+        $country = Country::BRAZIL;
+        $totalInstalments = 12;
+        $usdToBrlRate = 1;
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config());
+        $instalmentsArray = $creditCard->getInstalmentsByCountry($country);
+        $expectedInstalmentsArray = [];
+        for ($i = 1; $i <= 12; $i++) {
+            $expectedInstalmentsArray[$i] = $i;
+        }
+
+        $this->assertEquals($totalInstalments, count($instalmentsArray));
+        $this->assertEquals($expectedInstalmentsArray, $instalmentsArray);
+    }
+
+    public function testInstalmentsMexico()
+    {
+        $country = Country::MEXICO;
+        $totalInstalments = 5;
+        $usdToBrlRate = 1;
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config());
+        $instalmentsArray = $creditCard->getInstalmentsByCountry($country);
+        $expectedInstalmentsArray = array(
+            '1' => '1',
+            '3' => '3',
+            '6' => '6',
+            '9' => '9',
+            '12' => '12');
+
+        $this->assertEquals($totalInstalments, count($instalmentsArray));
+        $this->assertEquals($expectedInstalmentsArray, $instalmentsArray);
+    }
+
+    public function testInstalmentsColombia()
+    {
+        $country = Country::COLOMBIA;
+        $totalInstalments = 36;
+        $usdToBrlRate = 1;
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config());
+        $instalmentsArray = $creditCard->getInstalmentsByCountry($country);
+        $expectedInstalmentsArray = [];
+        for ($i = 1; $i <= 36; $i++) {
+            $expectedInstalmentsArray[$i] = $i;
+        }
+
+        $this->assertEquals($totalInstalments, count($instalmentsArray));
+        $this->assertEquals($expectedInstalmentsArray, $instalmentsArray);
+    }
+
+    public function testConfigMaxInstalments()
+    {
+        $country = Country::BRAZIL;
+        $usdToBrlRate = 1;
+        $creditCardConfig = new CreditCardConfig(['maxInstalments' => 1]);
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config(), $creditCardConfig);
+        $instalmentsByCountry = $creditCard->getInstalmentsByCountry($country);
+        $maxInstalmentsExpected = 1;
+
+        $this->assertEquals($maxInstalmentsExpected, count($instalmentsByCountry));
+    }
+
+    public function testMaxInstalmentsWhenIsZero()
+    {
+        $country = Country::BRAZIL;
+        $usdToBrlRate = 1;
+        $creditCardConfig = new CreditCardConfig(['maxInstalments' => 0]);
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config(), $creditCardConfig);
+        $instalmentsByCountry = $creditCard->getInstalmentsByCountry($country);
+        $maxInstalmentsExpected = 1;
+
+        $this->assertEquals($maxInstalmentsExpected, count($instalmentsByCountry));
+    }
+
+    public function testMaxInstalmentBiggerThanDefault()
+    {
+        $country = Country::BRAZIL;
+        $usdToBrlRate = 1;
+        $creditCardConfig = new CreditCardConfig(['maxInstalments' => 36]);
+        $creditCard = $this->setupGateway($usdToBrlRate, new Config(), $creditCardConfig);
+        $instalmentsByCountry = $creditCard->getInstalmentsByCountry($country);
+        $maxInstalmentsExpected = 12;
+
+        $this->assertEquals($maxInstalmentsExpected, count($instalmentsByCountry));
+    }
+
     private function setupGateway($usdToBrlRate, $config, $creditCardConfig = null)
     {
         $client = $this->getMockedClient($this->getExchangeRateSuccessfulResponseJsonWithRate($usdToBrlRate));